+++ openstack/autoscaling/v1/groups/delete.go
@@ -2,8 +2,7 @@ package groups
 
 import "github.com/opentelekomcloud/gophertelekomcloud"
 
-// Delete is a method of deleting a group by group id
-func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
-	_, r.Err = client.Delete(deleteURL(client, id), nil)
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("scaling_group", id), nil)
 	return
 }
