--- openstack/autoscaling/v1/groups/urls.go
@@ -1,31 +1,0 @@
-package groups
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-const resourcePath = "scaling_group"
-
-func createURL(c *golangsdk.ServiceClient) string {
-	return c.ServiceURL(resourcePath)
-}
-
-func deleteURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id)
-}
-
-func getURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id)
-}
-
-func listURL(c *golangsdk.ServiceClient) string {
-	return c.ServiceURL(resourcePath)
-}
-
-func enableURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id, "action")
-}
-
-func updateURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id)
-}
